+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 0.0.3 (15 Aug 2021)
+
+-   (Feat) Notify if it can be updated
+-   (Feat) Notify create cache on first run
+-   (Feat) Added argument `-f` to forcefully overwrite any current available `.gitignore` file, so generate will not overwrite by default
+
 ## 0.0.2 (14 Aug 2021)
 
 -   (Fix) README.md preview & slim package size